refactor(payments-v2): dedup — the e2e certifies the REAL wiring, one source for cross-repo protocol strings, shared test support
From the systematic duplication sweep (owner-prompted after catching the
paymentsV2 option-flag duplication):

- core/wallet-api-protocol.ts is now THE source for AUTH_CHALLENGE_PREFIX,
  completeSignMessage, progressSignMessage and verifyChallengeTemplate; the
  old wallet-api/ modules re-export it and impl/wallet-api-v2 deletes its
  re-implementations — the v2 challenge verifier had already drifted
  textually; the merged one keeps the strictest union of both (all tampering
  tests green unmodified).
- the vertical staging e2e imports authedWireClient/requestMemoCodec from
  core/payments-v2-wiring.ts instead of hand-copying them (an e2e that can
  drift from the code it certifies proves nothing); duplicated protocol
  comments deleted with the copies.
- tests/e2e/support/staging.ts + tests/unit/payments-v2/support.ts absorb
  8 in-memory KVs, 3 nodeWsFactories, 3 staging env/trustbase blocks,
  2 FakeSessions, and the small stub zoo; pubOf → core/crypto.
- tests/e2e/tmp-mint-probe.test.ts deleted (self-declared temp diagnostic,
  committed by mistake).

Net −324 lines. Suite 3324 green; 17/17 probes; live mint leg re-verified
against staging on the rewired e2e composition.
