+++ tests/unit/payments-v2/inventory.test.ts
@@ -6,7 +6,7 @@ import {
   type ReAddResult,
 } from '../../../modules/payments-v2/inventory/InventoryView';
 import type { InventoryItem, InventoryPage, StoragePort } from '../../../modules/payments-v2/ports';
-import type { ScopedKV } from '../../../modules/payments-v2/stores';
+import { deferred, memoryKV, type MemoryKV } from './support';
 
 const COIN = 'aa'.repeat(32);
 
@@ -32,32 +32,6 @@ function item(tokenId: string, opts: ItemOpts): InventoryItem {
 
 function page(items: InventoryItem[], cursor: number, more = false): InventoryPage {
   return { cursor, syncEpoch: 'e1', more, items };
-}
-
-function deferred<T>(): { promise: Promise<T>; resolve: (v: T) => void; reject: (e: unknown) => void } {
-  let resolve!: (v: T) => void;
-  let reject!: (e: unknown) => void;
-  const promise = new Promise<T>((res, rej) => {
-    resolve = res;
-    reject = rej;
-  });
-  return { promise, resolve, reject };
-}
-
-function makeKV(seed?: Record<string, unknown>): ScopedKV & { data: Map<string, unknown> } {
-  const data = new Map<string, unknown>(Object.entries(seed ?? {}));
-  return {
-    data,
-    async get<T>(key: string): Promise<T | null> {
-      return data.has(key) ? (data.get(key) as T) : null;
-    },
-    async set<T>(key: string, value: T): Promise<void> {
-      data.set(key, JSON.parse(JSON.stringify(value)));
-    },
-    async remove(key: string): Promise<void> {
-      data.delete(key);
-    },
-  };
 }
 
 function makePort(script: PageSource[]): {
@@ -96,12 +70,12 @@ function makeView(
   kvSeed?: Record<string, unknown>
 ): {
   view: InventoryView;
-  kv: ScopedKV & { data: Map<string, unknown> };
+  kv: MemoryKV;
   calls: (number | undefined)[];
   queue: PageSource[];
   events: string[];
 } {
-  const kv = makeKV(kvSeed);
+  const kv = memoryKV({ ...(kvSeed !== undefined ? { seed: kvSeed } : {}) });
   const { port, calls, queue } = makePort(script);
   const events: string[] = [];
   const view = new InventoryView({ port, kv, emit: (e) => events.push(e) });
@@ -161,12 +135,12 @@ describe('InventoryView pull protocol', () => {
   it('delta() resumes from the persisted §6 cursor record: since=<prior cursor>, then the ADVANCED cursor', async () => {
     const { view, kv, queue, calls } = makeView([page([item('A', { seq: 1 })], 5), page([], 5)]);
     await view.fullPull();
-    expect(kv.data.get('cursor:inventory')).toEqual({ cursor: 5, syncEpoch: 'e1' });
+    expect(kv.map.get('cursor:inventory')).toEqual({ cursor: 5, syncEpoch: 'e1' });
 
     queue.push(page([item('A', { seq: 8, state: 'S2' })], 8));
     await view.delta();
     expect(calls).toEqual([undefined, 5, 5]);
-    expect(kv.data.get('cursor:inventory')).toEqual({ cursor: 8, syncEpoch: 'e1' });
+    expect(kv.map.get('cursor:inventory')).toEqual({ cursor: 8, syncEpoch: 'e1' });
 
     queue.push(page([], 8));
     await view.delta();
@@ -190,7 +164,7 @@ describe('InventoryView pull protocol', () => {
 
     expect(calls).toEqual([undefined, 5, 5, undefined, 9]);
     expect(view.tokens(registry).map((t) => t.id)).toEqual(['B']);
-    expect(kv.data.get('cursor:inventory')).toEqual({ cursor: 9, syncEpoch: 'e2' });
+    expect(kv.map.get('cursor:inventory')).toEqual({ cursor: 9, syncEpoch: 'e2' });
   });
 
   it('emits inventory:updated only on net change — an identical re-delivered delta stays silent', async () => {
@@ -217,11 +191,11 @@ describe('InventoryView suspectedSpent overlay (F7)', () => {
     queue.push(() => gate.promise, page([], 6));
     const pull = view.fullPull();
     await view.demote('T', 'S1');
-    expect(kv.data.get('suspected-spent')).toEqual(['T:S1']);
+    expect(kv.map.get('suspected-spent')).toEqual(['T:S1']);
     gate.resolve(page([item('T', { seq: 1 }), item('U', { seq: 2, amount: '50' })], 6));
     await pull;
     expect(view.pool(COIN).map((p) => p.tokenId)).toEqual(['U']);
-    expect(kv.data.get('suspected-spent')).toEqual(['T:S1']);
+    expect(kv.map.get('suspected-spent')).toEqual(['T:S1']);
     const tokens = view.tokens(registry);
     expect(tokens.find((t) => t.id === 'T')?.suspectedSpent).toBe(true);
     expect(tokens.find((t) => t.id === 'U')?.suspectedSpent).toBeUndefined();
@@ -238,7 +212,7 @@ describe('InventoryView suspectedSpent overlay (F7)', () => {
     );
     await view.fullPull();
     expect(view.pool(COIN).map((p) => p.tokenId)).toEqual(['U']);
-    expect(kv.data.get('suspected-spent')).toEqual(['T:S1']);
+    expect(kv.map.get('suspected-spent')).toEqual(['T:S1']);
   });
 
   it('a state advance is positive evidence: the stale demotion prunes and the token returns to the pool', async () => {
@@ -250,7 +224,7 @@ describe('InventoryView suspectedSpent overlay (F7)', () => {
     await view.delta();
     expect(view.pool(COIN)).toEqual([{ tokenId: 'T', amount: 100n }]);
     await vi.waitFor(() => {
-      expect(kv.data.get('suspected-spent')).toEqual([]);
+      expect(kv.map.get('suspected-spent')).toEqual([]);
     });
   });
 
@@ -277,7 +251,7 @@ describe('InventoryView recoverRemoved (F6)', () => {
     );
     expect(r1.confirmedSpent).toEqual([{ tokenId: 'T', stateHash: 'S1' }]);
     expect(r1.recovered).toEqual([]);
-    expect(kv.data.get('known-spends')).toEqual(['T:S1']);
+    expect(kv.map.get('known-spends')).toEqual(['T:S1']);
     expect(view.pool(COIN)).toEqual([]);
 
     const r1b = await view.recoverRemoved(noRecoverCalls.getBlobs, noRecoverCalls.isSpent, noRecoverCalls.reAdd);
@@ -301,7 +275,7 @@ describe('InventoryView recoverRemoved (F6)', () => {
     const r = await view.recoverRemoved(getBlobs, async () => true, reAdd);
     expect(r.confirmedSpent).toEqual([{ tokenId: 'T', stateHash: 'S1' }]);
     expect(reAdd).not.toHaveBeenCalled();
-    expect(kv.data.get('known-spends')).toEqual(['T:S1']);
+    expect(kv.map.get('known-spends')).toEqual(['T:S1']);
     expect(view.tokens(registry)).toEqual([]);
     await view.recoverRemoved(getBlobs, async () => true, reAdd);
     expect(getBlobs).toHaveBeenCalledTimes(1);
@@ -328,17 +302,17 @@ describe('InventoryView recoverRemoved (F6)', () => {
       confirmedSpent: [],
     });
     expect(getBlobs).not.toHaveBeenCalled();
-    expect(kv.data.has('known-spends')).toBe(false);
+    expect(kv.map.has('known-spends')).toBe(false);
 
     await view.fullPull();
     await view.recoverRemoved(getBlobs, isSpent, reAdd);
     expect(getBlobs).toHaveBeenCalledTimes(1);
-    expect(kv.data.get('known-spends')).toEqual(['T:S1']);
+    expect(kv.map.get('known-spends')).toEqual(['T:S1']);
     expect(queue).toHaveLength(0);
   });
 
   it('a recorded knownSpend survives a reload: a fresh view over the same KV never re-probes that state', async () => {
-    const kv = makeKV();
+    const kv = memoryKV();
     const script = (): PageSource[] => [page([item('T', { seq: 5, status: 'removed' })], 5), page([], 5)];
     const first = makePort(script());
     const view1 = new InventoryView({ port: first.port, kv, emit: () => undefined });
@@ -348,7 +322,7 @@ describe('InventoryView recoverRemoved (F6)', () => {
       async () => true,
       async (): Promise<ReAddResult> => 'added'
     );
-    expect(kv.data.get('known-spends')).toEqual(['T:S1']);
+    expect(kv.map.get('known-spends')).toEqual(['T:S1']);
 
     const second = makePort(script());
     const view2 = new InventoryView({ port: second.port, kv, emit: () => undefined });
